Update to 0.2.10.5

+++ src/com/lilithsthrone/game/character/body/valueEnums/Height.java
@@ -67,7 +67,7 @@ public enum Height {
 	/**
 	 * If height is less than this value, then that height is short stature.
 	 */
-	public static int getShortStatureCutOff() {
+	public static float getShortStatureCutOff() {
 		return ZERO_TINY.getMinimumValue();
 	}
